+++ TIUIElements/Sources/Appearance/UIVIew+AppearanceConfigurable.swift
@@ -20,14 +20,24 @@
 //  THE SOFTWARE.
 //
 
+import TIUIKitCore
 import UIKit
 
-extension UILabel {
-    public func configureUILabel<L: ViewLayout>(appearance: BaseAppearance<L>) {
-        appearance.textAttributes?
-            .configure(label: self,
-                       with: attributedText?.string ?? text)
+extension UIView {
+    public func configureUIView(appearance: BaseAppearance<some ViewLayout>) {
+        backgroundColor = appearance.backgroundColor
+        layer.masksToBounds = true
+        layer.maskedCorners = appearance.roundedCorners
+        layer.cornerRadius = appearance.cornerRadius
 
-        super.configureUIView(appearance: appearance)
+        guard let shadow = appearance.shadow else {
+            return
+        }
+
+        layer.shadowOpacity = shadow.opacity
+        layer.shadowOffset = shadow.offset
+        layer.shadowColor = shadow.color.cgColor
+        layer.shadowRadius = shadow.radius
+        clipsToBounds = false
     }
 }
