+++ TIUIElements/Sources/Appearance/UIView+Appearance.swift
@@ -20,6 +20,7 @@
 //  THE SOFTWARE.
 //
 
+import TIUIKitCore
 import UIKit
 
 extension UIView {
