+++ TIUIKitCore/TIUIKitCore.podspec
@@ -1,10 +1,11 @@
 Pod::Spec.new do |s|
   s.name             = 'TIUIKitCore'
-  s.version          = '1.32.1'
+  s.version          = '1.33.0'
   s.summary          = 'Core UI elements: protocols, views and helpers.'
   s.homepage         = 'https://github.com/TouchInstinct/LeadKit/tree/' + s.version.to_s + '/' + s.name
   s.license          = { :type => 'MIT', :file => 'LICENSE' }
-  s.author           = { 'petropavel13' => 'ivan.smolin@touchin.ru' }
+  s.author           = { 'petropavel13' => 'ivan.smolin@touchin.ru',
+                         'castlele' => 'nikita.semenov@touchin.ru' }
   s.source           = { :git => 'https://github.com/TouchInstinct/LeadKit.git', :tag => s.version.to_s }
 
   s.ios.deployment_target = '11.0'
